requirements.txt
# Core Dependencies
openai>=1.0.0
pydantic>=2.12.0
neo4j>=5.14.0
pyyaml>=6.0
python-dotenv>=1.0.0

# Agent Framework
langgraph>=0.0.20
langchain>=0.1.0
langchain-openai>=0.0.5

# LlamaIndex for RAG and Atomic Tasks
llama-index>=0.9.0
llama-index-llms-openai>=0.1.0
llama-index-embeddings-openai>=0.1.0

# Document Processing
PyPDF2>=3.0.0
pdfplumber>=0.9.0
python-docx>=1.0.0
pypdf>=3.17.0
PyMuPDF>=1.23.0
Pillow>=10.0.0

# Vector Search & Embeddings (FAISS)
faiss-cpu>=1.7.4
numpy>=1.24.0
scikit-learn>=1.3.0

# CLI & UI
rich>=13.0.0
prompt-toolkit>=3.0.0
gradio>=4.0.0

# API Server
fastapi>=0.109.0
uvicorn>=0.27.0
websockets>=12.0

# Web Browsing (Separate Browser Agent)
playwright>=1.40.0
# langextract>=0.5.0  # Optional - package not available on PyPI, code has fallback

# Utilities
tqdm>=4.65.0
# pathlib>=1.0.1  # Built-in module, no need to install
tiktoken>=0.5.0
yfinance>=0.2.28
python-dateutil>=2.8.0
packaging>=24.0
google-api-python-client>=2.153.0
youtube-transcript-api>=0.6.2

# macOS Window Management
pyobjc-framework-Quartz>=9.0

# HTTP Clients
requests>=2.31.0
requests-oauthlib>=1.3.1
httpx[http2]>=0.25.0  # For OpenAI connection pooling with HTTP/2 support
slack_sdk>=3.26.2

# Persistence
motor>=3.4.0

# Google Search (no API key required)
googlesearch-python>=1.3.0

# Google Maps API (optional - for maps functionality)
googlemaps>=4.10.0

# YAML handling (optional - for better config preservation)
ruamel.yaml>=0.18.0

# Testing
pytest>=7.4.0
pytest-timeout>=2.2.0

# Telemetry & Observability
opentelemetry-distro>=0.44b0
opentelemetry-exporter-otlp-proto-grpc>=1.23.0
opentelemetry-instrumentation>=0.44b0
opentelemetry-instrumentation-fastapi>=0.44b0
opentelemetry-instrumentation-httpx>=0.44b0
opentelemetry-sdk>=1.23.0
